モーダル周りを整形
